Removed old view and edit components, added progress spinner to dataset details dashboard, and updated unit tests
