+++ src/app/datasets/dataset-details-dashboard/dataset-details-dashboard.component.html
@@ -116,3 +116,7 @@
     </mat-card>
   </mat-tab>
 </mat-tab-group>
+
+<div *ngIf="loading$ | async" class="spinner">
+  <mat-spinner></mat-spinner>
+</div>
