--- src/app/datasets/dataset-form/dataset-form.component.scss
@@ -1,55 +1,0 @@
-.container {
-  margin: 1em;
-
-  button {
-    margin: 1em 1em 0 0;
-  }
-
-  .material-icons {
-    vertical-align: middle;
-  }
-
-  .formRow {
-    .deleteButton {
-      color: #c81919;
-    }
-  }
-
-  .addButton {
-    color: #ffffff;
-    background: #0ea432;
-  }
-
-  .submitButton {
-    color: #ffffff;
-    background: #0094ca;
-  }
-}
-
-@media only screen and (min-width: 1280px) {
-  .container {
-    .formRow {
-      .formColumn {
-        padding-right: 1em;
-
-        .selectField {
-          max-width: 8em;
-        }
-      }
-    }
-  }
-}
-
-@media only screen and (max-width: 1279px) {
-  .container {
-    .formRow {
-      .formColumn {
-        padding-right: none;
-
-        .selectField {
-          max-width: none;
-        }
-      }
-    }
-  }
-}
