--- src/app/datasets/metadata-table/metadata-table.component.scss
@@ -1,23 +1,0 @@
-.metadataTable {
-  max-width: 50em;
-
-  .mat-column-name {
-    flex: 1 0 7em;
-  }
-
-  .mat-column-value {
-    flex: 1 0 7em;
-  }
-
-  .mat-column-unit {
-    flex: 1 0 5em;
-  }
-}
-
-@media only screen and (max-width: 1279px) {
-  .metadataTable {
-    .mat-column-unit {
-      justify-content: right !important;
-    }
-  }
-}
